requirements.txt
PySide6-Fluent-Widgets
pymongo[srv]
boto3
